Add the folke/noice nvim plugin
It removes the bottom bar and makes the command line be in a popup in
the middle of the screen instead of the bottom. It also shows messages
in the bottom right. It might duplicate some of the stuff that fidget
does, so I might be able to remove that.