Make foldlevel better
When opening a file the initial file has a foldlevel of 0 and that
causes all folds to be folded. Also foldlevel of 99 means that using
things like zm don't really work. Using an autocmd to call zR on the
buffer sets the foldlevel to the max foldlevel for the document, so that
everything is unfolded and you can still use zm to reduce the level for
the whole buffer.