Add command to quickly install all packages for a type of dev environment
Currently when I start a codespace I have to type out all the packages
that I want to install. It isn't that much work, but remembering all of
them is a bit annoying and I tend to forget one or two. This accepts the
type of environment as the first arg or try to detect it based on the
files in the directory and then installs a pre-configured set of
packages.