Make a progress indicator per linter
I was making one progress indicator for all linters that I don't think
actually worked. This makes one per linter all under nvim-lint client
and then has a timer to check if they are still running or not. I didn't
test it a ton, but it seemed to work in the little bit of testing I did.