Create a key bind for opening a terminal in a float
I want to do some stuff real quick in the terminal and it is nice to
have quick access to one in a floating window. This will let me do my
thing real quick and then close the float. I am keeping the same
terminal session in case I want to continue where I left off and there
isn't much of a downside. I can always ctrl-d the session to and reopen
the float to get a new shell.